resume_analyzer_agent/requirements.txt
google-adk>=0.1.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0